Fixed Android and iOS specific issues

+++ base/src/iosMain/kotlin/utils/Decimal.kt
@@ -150,14 +150,14 @@ actual fun Decimal.times(value: Decimal, scale: Int, roundingMode: RoundingMode)
 actual infix fun Decimal.pow(n: Int): Decimal = if (n < 0)
     1.toDecimal() / pow(n.absoluteValue)
 else
-    copy(nsDecimal = nsDecimal.decimalNumberByRaisingToPower(n.toULong()))
+    copy(nsDecimal = nsDecimal.decimalNumberByRaisingToPower(n.toULong() as NSUInteger))
 
 actual fun Decimal.pow(n: Int, scale: Int): Decimal = if (n < 0)
     1.toDecimal() / pow(n.absoluteValue, scale)
 else
     copy(
         nsDecimal = nsDecimal.decimalNumberByRaisingToPower(
-            n.toULong(),
+            n.toULong() as NSUInteger,
             NSDecimalNumberHandler(
                 roundingMode = NSRoundingMode.NSRoundPlain,
                 scale = scale.toShort(),
